[Update] move AdminLayout components into main folder
