+++ frontend/src/components/NavbarMinimise/NavbarMinimise.tsx
