+++ frontend/src/components/NavbarMinimise/index.ts
