+++ frontend/src/components/SidebarContent/index.ts
