+++ frontend/src/components/SidebarLogo/SidebarLogo.tsx
