+++ frontend/src/components/SidebarLogo/index.ts
